# web/gen/types/Badge.ts
export type Badge = {
    /**
     * @type string | undefined
    */
    color?: string;
    /**
     * @type string | undefined
    */
    created_at?: string;
    /**
     * @type string | undefined
    */
    icon?: string;
    /**
     * @type integer | undefined
    */
    id?: number;
    /**
     * @type string | undefined
    */
    text?: string;
};